Rename dendrite-p2p-demo -> dendrite-demo-libp2p
